Use the built in IoC, faster tests..
