--- src/test/datascience/kernelLauncher.functional.test.ts
@@ -1,192 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-'use strict';
-
-import { assert, use } from 'chai';
-
-import { KernelMessage } from '@jupyterlab/services';
-import * as uuid from 'uuid/v4';
-import { IProcessServiceFactory } from '../../client/common/process/types';
-import { createDeferred } from '../../client/common/utils/async';
-import { JupyterZMQBinariesNotFoundError } from '../../client/datascience/jupyter/jupyterZMQBinariesNotFoundError';
-import { KernelDaemonPool } from '../../client/datascience/kernel-launcher/kernelDaemonPool';
-import { KernelLauncher } from '../../client/datascience/kernel-launcher/kernelLauncher';
-import { IKernelConnection, IKernelFinder } from '../../client/datascience/kernel-launcher/types';
-import { createRawKernel } from '../../client/datascience/raw-kernel/rawKernel';
-import { IJupyterKernelSpec, IKernelDependencyService } from '../../client/datascience/types';
-import { PythonEnvironment } from '../../client/pythonEnvironments/info';
-import { sleep, waitForCondition } from '../common';
-import { DataScienceIocContainer } from './dataScienceIocContainer';
-import { takeSnapshot, writeDiffSnapshot } from './helpers';
-import { MockKernelFinder } from './mockKernelFinder';
-import { requestExecute } from './raw-kernel/rawKernelTestHelpers';
-
-// Chai as promised is not part of this file
-import * as chaiAsPromised from 'chai-as-promised';
-import { IPythonExtensionChecker } from '../../client/api/types';
-import { IFileSystem } from '../../client/common/platform/types';
-import { KernelEnvironmentVariablesService } from '../../client/datascience/kernel-launcher/kernelEnvVarsService';
-import { traceInfo } from '../../client/common/logger';
-use(chaiAsPromised);
-
-const test_Timeout = 30_000;
-
-suite('DataScience - Kernel Launcher', () => {
-    let ioc: DataScienceIocContainer;
-    let kernelLauncher: KernelLauncher;
-    let pythonInterpreter: PythonEnvironment | undefined;
-    let kernelSpec: IJupyterKernelSpec;
-    let kernelFinder: MockKernelFinder;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    let snapshot: any;
-
-    suiteSetup(() => {
-        snapshot = takeSnapshot();
-    });
-
-    setup(async function () {
-        traceInfo(`Start Test ${this.currentTest?.title}`);
-        ioc = new DataScienceIocContainer();
-        ioc.registerDataScienceTypes();
-        kernelFinder = new MockKernelFinder(ioc.get<IKernelFinder>(IKernelFinder));
-        const processServiceFactory = ioc.get<IProcessServiceFactory>(IProcessServiceFactory);
-        const daemonPool = ioc.get<KernelDaemonPool>(KernelDaemonPool);
-        const fileSystem = ioc.get<IFileSystem>(IFileSystem);
-        const extensionChecker = ioc.get<IPythonExtensionChecker>(IPythonExtensionChecker);
-        kernelLauncher = new KernelLauncher(
-            processServiceFactory,
-            fileSystem,
-            daemonPool,
-            extensionChecker,
-            ioc.get<KernelEnvironmentVariablesService>(KernelEnvironmentVariablesService),
-            ioc.get<IKernelDependencyService>(IKernelDependencyService)
-        );
-
-        await ioc.activate();
-        if (!ioc.mockJupyter) {
-            pythonInterpreter = await ioc.getJupyterCapableInterpreter();
-            kernelSpec = {
-                argv: [pythonInterpreter!.path, '-m', 'ipykernel_launcher', '-f', '{connection_file}'],
-                display_name: 'new kernel',
-                language: 'python',
-                name: 'newkernel',
-                path: 'path',
-                env: undefined
-            };
-        }
-        traceInfo(`Start Test Complete ${this.currentTest?.title}`);
-    });
-    teardown(function () {
-        traceInfo(`End Test ${this.currentTest?.title}`);
-    });
-    suiteTeardown(() => {
-        writeDiffSnapshot(snapshot, 'KernelLauncher');
-    });
-
-    test('Launch from kernelspec', async function () {
-        if (!process.env.VSC_FORCE_REAL_JUPYTER) {
-            // eslint-disable-next-line no-invalid-this
-            this.skip();
-        } else {
-            let exitExpected = false;
-            const deferred = createDeferred<boolean>();
-            const kernel = await kernelLauncher.launch(
-                { kernelSpec, kind: 'startUsingKernelSpec' },
-                -1,
-                undefined,
-                process.cwd()
-            );
-            kernel.exited(() => {
-                if (exitExpected) {
-                    deferred.resolve(true);
-                } else {
-                    deferred.reject(new Error('Kernel exited prematurely'));
-                }
-            });
-
-            assert.isOk<IKernelConnection | undefined>(kernel.connection, 'Connection not found');
-
-            // It should not exit.
-            await assert.isRejected(
-                waitForCondition(() => deferred.promise, 2_000, 'Timeout'),
-                'Timeout'
-            );
-
-            // Upon disposing, we should get an exit event within 100ms or less.
-            // If this happens, then we know a process existed.
-            exitExpected = true;
-            await kernel.dispose();
-            await deferred.promise;
-        }
-    }).timeout(test_Timeout);
-
-    test('Launch with environment', async function () {
-        if (!process.env.VSC_FORCE_REAL_JUPYTER || !pythonInterpreter) {
-            // eslint-disable-next-line no-invalid-this
-            this.skip();
-        } else {
-            const spec: IJupyterKernelSpec = {
-                name: 'foo',
-                language: 'python',
-                path: pythonInterpreter.path,
-                display_name: pythonInterpreter.displayName || 'foo',
-                argv: [pythonInterpreter.path, '-m', 'ipykernel_launcher', '-f', '{connection_file}'],
-                env: {
-                    TEST_VAR: '1'
-                }
-            };
-            kernelFinder.addKernelSpec(pythonInterpreter.path, spec);
-
-            const kernel = await kernelLauncher.launch(
-                { kernelSpec: spec, kind: 'startUsingKernelSpec' },
-                30_000,
-                undefined,
-                process.cwd()
-            );
-
-            assert.isOk<IKernelConnection | undefined>(kernel.connection, 'Connection not found');
-
-            // Send a request to print out the env vars
-            const rawKernel = createRawKernel(kernel, uuid());
-
-            const result = await requestExecute(rawKernel, 'import os\nprint(os.getenv("TEST_VAR"))');
-            assert.ok(result, 'No result returned');
-            // Should have a stream output message
-            const output = result.find((r) => r.header.msg_type === 'stream') as KernelMessage.IStreamMsg;
-            assert.ok(output, 'no stream output');
-            assert.equal(output.content.text, '1\n', 'Wrong content found on message');
-
-            // Upon disposing, we should get an exit event within 100ms or less.
-            // If this happens, then we know a process existed.
-            await kernel.dispose();
-        }
-    }).timeout(test_Timeout);
-
-    test('Bind with ZMQ', async function () {
-        if (!process.env.VSC_FORCE_REAL_JUPYTER) {
-            // eslint-disable-next-line no-invalid-this
-            this.skip();
-        } else {
-            const kernel = await kernelLauncher.launch(
-                { kernelSpec, kind: 'startUsingKernelSpec' },
-                -1,
-                undefined,
-                process.cwd()
-            );
-
-            try {
-                const zmq = await import('zeromq');
-                const sock = new zmq.Pull();
-
-                sock.connect(`tcp://${kernel.connection!.ip}:${kernel.connection!.stdin_port}`);
-                sock.receive().ignoreErrors(); // This will never return unless the kenrel process sends something. Just used for testing the API is available
-                await sleep(50);
-                sock.close();
-            } catch (e) {
-                throw new JupyterZMQBinariesNotFoundError(e.toString());
-            } finally {
-                await kernel.dispose();
-            }
-        }
-    });
-});
